Refactor fetchLogFile and extract stream decode functionality
